feat(front): refactor file structure and import paths
