+++ front/src/layout/Footer.svelte
