+++ front/src/layout/NavBar.svelte
