+++ front/src/layout/NavHome.svelte
