+++ front/src/styles/components/Code/Code.css
@@ -1,3 +1,4 @@
+@import "syntax.css";
 .syntax-highlighter {
   @apply shadow-lg
     rounded-lg
