+++ front/src/styles/components/fonts.css
