+++ front/src/styles/components/loading.css
