+++ front/src/styles/pages/not-found.css
@@ -1,5 +1,5 @@
-@import "./components.css";
-@import "./utilities.css";
+@import "../components/tw-components.css";
+@import "../components/utilities.css";
 @tailwind components;
 @tailwind utilities;
