+++ front/src/styles/pages/posts.css
@@ -1,5 +1,5 @@
-@import "./components.css";
-@import "./utilities.css";
+@import "../components/tw-components.css";
+@import "../components/utilities.css";
 
 .posts,
 .post,
